requirements.txt
annotated-types==0.6.0
anyio==4.3.0
authlib==1.3.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
fastapi==0.110.1
fpdf==1.7.2
google-auth==2.30.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.1
itsdangerous==2.2.0
lazy_loader==0.4
networkx==3.3
numpy==1.26.4
openai==1.16.2
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
scikit-image==0.23.2
scipy==1.13.0
sniffio==1.3.1
starlette==0.37.2
tifffile==2024.5.10
tqdm==4.66.2
typing_extensions==4.11.0
gunicorn
uvicorn==0.27.1
python-jose==3.3.0
werkzeug==3.0.4
